+++ packages/react/CHANGELOG.md
@@ -1,5 +1,54 @@
 # Changelog
 
+## [1.125.0](https://github.com/factorialco/factorial-one/compare/factorial-one-react-v1.124.0...factorial-one-react-v1.125.0) (2025-07-14)
+
+
+### Features
+
+* datacollection grouping on nested values ([#2240](https://github.com/factorialco/factorial-one/issues/2240)) ([6855042](https://github.com/factorialco/factorial-one/commit/6855042efd4fff42ae3ae1201400a13f5aba83b6))
+
+## [1.124.0](https://github.com/factorialco/factorial-one/compare/factorial-one-react-v1.123.0...factorial-one-react-v1.124.0) (2025-07-11)
+
+
+### Features
+
+* **Card:** make Emoji avatar available ([#2248](https://github.com/factorialco/factorial-one/issues/2248)) ([7dba9fd](https://github.com/factorialco/factorial-one/commit/7dba9fdf65ae3403b6c9c7c5b6e2232bef72d0c6))
+
+## [1.123.0](https://github.com/factorialco/factorial-one/compare/factorial-one-react-v1.122.2...factorial-one-react-v1.123.0) (2025-07-11)
+
+
+### Features
+
+* Use ellipsis in icon names in Icons documentation ([#2246](https://github.com/factorialco/factorial-one/issues/2246)) ([d3aaa89](https://github.com/factorialco/factorial-one/commit/d3aaa89b612ecbbf21d10f81500fcb02d08827af))
+
+## [1.122.2](https://github.com/factorialco/factorial-one/compare/factorial-one-react-v1.122.1...factorial-one-react-v1.122.2) (2025-07-10)
+
+
+### Bug Fixes
+
+* stop using module icon prop in breadcrumbs ([#2233](https://github.com/factorialco/factorial-one/issues/2233)) ([8c325f5](https://github.com/factorialco/factorial-one/commit/8c325f56ebf9c83d70e1e223906e805af57218bf))
+
+## [1.122.1](https://github.com/factorialco/factorial-one/compare/factorial-one-react-v1.122.0...factorial-one-react-v1.122.1) (2025-07-10)
+
+
+### Bug Fixes
+
+* **datacollection:** filters form width and custo height by type ([#2160](https://github.com/factorialco/factorial-one/issues/2160)) ([6c067f2](https://github.com/factorialco/factorial-one/commit/6c067f21b04f0d24f8b2cf8f73c01bb219b19788))
+
+## [1.122.0](https://github.com/factorialco/factorial-one/compare/factorial-one-react-v1.121.1...factorial-one-react-v1.122.0) (2025-07-10)
+
+
+### Features
+
+* **Icon:** add color prop ([#2239](https://github.com/factorialco/factorial-one/issues/2239)) ([2120ad8](https://github.com/factorialco/factorial-one/commit/2120ad841992fde01276d8309c3406b87129e9a0))
+
+## [1.121.1](https://github.com/factorialco/factorial-one/compare/factorial-one-react-v1.121.0...factorial-one-react-v1.121.1) (2025-07-10)
+
+
+### Bug Fixes
+
+* proper overflow calculation to avoid overflows ([#2238](https://github.com/factorialco/factorial-one/issues/2238)) ([2c86bc5](https://github.com/factorialco/factorial-one/commit/2c86bc5107d400faf6826c7f22d4c6f6a7c68c9a))
+
 ## [1.121.0](https://github.com/factorialco/factorial-one/compare/factorial-one-react-v1.120.0...factorial-one-react-v1.121.0) (2025-07-10)